requirements.txt
beautifulsoup4==4.6.0
bitstring==3.1.5
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
click==6.7
DateTime==4.2
Django==1.11.7
gnewsclient==1.0.2
google-api-python-client==1.6.4
googlefinance.client==1.3.0
googlemaps==2.5.1
holidays==0.8.1
html5lib==0.999999999
httplib2==0.10.3
idna==2.6
keyring==8.7
keyrings.alt==1.3
numpy==1.13.3
oauth2client==4.1.2
pandas==0.21.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pyicloud==0.9.1
pyowm==2.7.1
python-dateutil==2.6.1
pytz==2017.3
pywin32-ctypes==0.1.2
requests==2.18.4
rsa==3.4.2
six==1.11.0
SpeechRecognition==3.7.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
zope.interface==4.4.3
